Refactoring) Fixture를 나누기
- Tasks 와 관련된 Fixture 와 화면에 보이는 UI 관련 Fixture를 나누었다.
