+++ src/Fixture/Tasks.js
@@ -5,20 +5,11 @@ const TASKS = [
   { id: 3, title: '운동하기' },
 ];
 const TASK = { id: 100, title: 'coodesoom 과제 제출하기' };
-
 const TASK_TITLE = '자바스크립트 공부하기';
-const PLACEHOLDER = '할 일을 입력해 주세요';
-const ADDTASK_TEXT = '추가';
-const COMPLETE_TEXT = '완료';
-const EMPTY_TASK_TEXT = '할 일이 없어요!';
 
 export {
   EMPTY_TASKS,
   TASKS,
   TASK,
   TASK_TITLE,
-  PLACEHOLDER,
-  ADDTASK_TEXT,
-  COMPLETE_TEXT,
-  EMPTY_TASK_TEXT,
 };
