+++ src/Fixture/UserInterfaceText.js
@@ -1,0 +1,11 @@
+const PLACEHOLDER = '할 일을 입력해 주세요';
+const ADDTASK_TEXT = '추가';
+const COMPLETE_TEXT = '완료';
+const EMPTY_TASK_TEXT = '할 일이 없어요!';
+
+export {
+  PLACEHOLDER,
+  ADDTASK_TEXT,
+  COMPLETE_TEXT,
+  EMPTY_TASK_TEXT,
+};
