+++ src/Item.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
 
+import { COMPLETE_TEXT } from './Fixture/UserInterfaceText';
+
 export default function Item({ task: { id, title }, onClickDelete }) {
   return (
     <li>
       {title}
       <button type="button" onClick={() => onClickDelete(id)}>
-        완료
+        {COMPLETE_TEXT}
       </button>
     </li>
   );
