+++ src/List.jsx
@@ -2,10 +2,12 @@ import React from 'react';
 
 import Item from './Item';
 
+import { EMPTY_TASK_TEXT } from './Fixture/UserInterfaceText';
+
 export default function List({ tasks, onClickDelete }) {
   if (tasks.length === 0) {
     return (
-      <p>할 일이 없어요!</p>
+      <p>{EMPTY_TASK_TEXT}</p>
     );
   }
